191007 | restful & faker & gihpy

+++ 03_Django/04_django_crud_review/articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.core.exceptions import ValidationError
 from .models import Article
+from IPython import embed
 
 def index(request):
     # articles = Article.objects.all()[::-1]
@@ -10,20 +11,17 @@ def index(request):
     context = {'articles': articles}
     return render(request, 'articles/index.html', context)
 
-def new(request):
-    return render(request, 'articles/new.html')
-
 def create(request):
-    try:
+    # POST 요청일 때
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
         article = Article(title=title, content=content)
-        article.full_clean()
-    except ValidationError:
-        raise ValidationError('Your Error Message')
-    else:
         article.save()
         return redirect('articles:detail', article.pk)
+    # GET 요청일 때
+    else:
+        return render(request, 'articles/create.html')
     """
     title = request.POST.get('title')
     # print(title)
@@ -47,21 +45,25 @@ def create(request):
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
     context = {'article': article}
+    # embed()
     return render(request, 'articles/detail.html', context)
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
-
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {'article': article}
-    return render(request, 'articles/edit.html', context)
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
 def update(request, pk):
+    # embed()
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect('articles:detail', article.pk)
+    if request.method == 'POST':
+        article = Article.objects.get(pk=pk)
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.pk)
+    else:
+        context = {'article': article}
+        return render(request, 'articles/update.html', context)
